+++ src/main/java/com/example/demo/ECommerce/ERepositories/OfferRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OfferRepository extends JpaRepository<Offer, Long> {
-    List<Offer> findByProductId(Long productId);
-    List<Offer> findByBuyerId(Long buyerId);
+public interface OfferRepository extends JpaRepository<Offer, Integer> {
+    List<Offer> findByProductId(Integer productId);
+    List<Offer> findByBuyerId(Integer buyerId);
 }
