+++ src/main/java/com/example/demo/Entities/Complaint.java
@@ -13,7 +13,7 @@ public class Complaint {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String title;
 
@@ -80,11 +80,11 @@ public class Complaint {
         this.complainType = complainType;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
